backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
celery
redis
python-multipart
python-jose[cryptography]
bcrypt
emails
python-dotenv
httpx
pydantic
pydantic-settings
email-validator
pyotp
qrcode[pil]
groq
google-generativeai
tavily-python
